f from log

f from log

f_from_log(meta)

Arguments

meta
a data frame containing the log metadata for each set of alpha vectors desired, see meta_from_log

Value

the growth function associated with the model indicated.

Details

note this function is unique to the fisheries example problem and assumes that sarsop call is run with logging specifying a column "model" that contains either the string "ricker" (corresponding to a Ricker-type growth function) or "allen" (corresponding to an Allen-type.)

Examples

not_run({

  source(system.file("examples/fisheries-ex.R", package = "sarsop"))
  log = tempfile()
  alpha <- sarsop(transition, observation, reward, discount, precision = 10,
                  log_dir = log, log_data = log_data)

  ## Get metadata for all logged solutions matching the desired query
  meta <- meta_from_log(parameters = data.frame(model = "ricker", r = 0.1), log_dir = log)
  alphas <- alphas_from_log(meta, log_dir = log)
  fs <- f_from_log(meta)
  models <- models_from_log(meta)
})